decant admin commands
