+++ .env.example
@@ -1,5 +1,8 @@
 # .env
 
+# Comma-separated list of roles which can run admin commands.
+ADMIN_ROLES=''
+
 # Bot secret Discord token.
 DISCORD_TOKEN = ''
