+++ pombot/commands/__init__.py
@@ -1,7 +1,10 @@
+from pombot.commands.create_event import *
+from pombot.commands.events import *
+from pombot.commands.howmany import *
+from pombot.commands.newleaf import *
 from pombot.commands.pom import *
 from pombot.commands.poms import *
-from pombot.commands.howmany import *
-from pombot.commands.undo import *
-from pombot.commands.newleaf import *
+from pombot.commands.remove_event import *
 from pombot.commands.reset import *
-from pombot.commands.events import *
+from pombot.commands.total import *
+from pombot.commands.undo import *
